+++ konva.js
@@ -1,8 +1,8 @@
 /*
- * Konva JavaScript Framework v1.6.7
+ * Konva JavaScript Framework v1.7.2
  * http://konvajs.github.io/
  * Licensed under the MIT or GPL Version 2 licenses.
- * Date: Mon Jul 31 2017
+ * Date: Wed Oct 11 2017
  *
  * Original work Copyright (C) 2011 - 2013 by Eric Rowell (KineticJS)
  * Modified work Copyright (C) 2014 - 2017 by Anton Lavrenov (Konva)
@@ -28,7 +28,7 @@
  */
 
 // runtime check for already included Konva
-(function(global) {
+(function() {
   'use strict';
   /**
      * @namespace Konva
@@ -38,7 +38,7 @@
 
   var Konva = {
     // public
-    version: '1.6.7',
+    version: '1.7.2',
 
     // private
     stages: [],
@@ -48,6 +48,10 @@
     shapes: {},
     listenClickTap: false,
     inDblClickWindow: false,
+
+    isBrowser:
+      typeof window !== 'undefined' &&
+      {}.toString.call(window) === '[object Window]',
 
     // configurations
     enableTrace: false,
@@ -65,14 +69,14 @@
     pixelRatio: undefined,
     /**
          * Drag distance property. If you start to drag a node you may want to wait until pointer is moved to some distance from start point,
-         * only then start dragging.
+         * only then start dragging. Default is 3px.
          * @property dragDistance
          * @default 0
          * @memberof Konva
          * @example
          * Konva.dragDistance = 10;
          */
-    dragDistance: 0,
+    dragDistance: 3,
     /**
          * Use degree values for angle properties. You may set this property to false if you want to use radiant values.
          * @property angleDeg
@@ -196,12 +200,14 @@
     _parseUA: function(userAgent) {
       var ua = userAgent.toLowerCase(),
         // jQuery UA regex
-        match = /(chrome)[ \/]([\w.]+)/.exec(ua) ||
-        /(webkit)[ \/]([\w.]+)/.exec(ua) ||
-        /(opera)(?:.*version|)[ \/]([\w.]+)/.exec(ua) ||
-        /(msie) ([\w.]+)/.exec(ua) ||
-        (ua.indexOf('compatible') < 0 &&
-          /(mozilla)(?:.*? rv:([\w.]+)|)/.exec(ua)) || [],
+        match =
+          /(chrome)[ /]([\w.]+)/.exec(ua) ||
+          /(webkit)[ /]([\w.]+)/.exec(ua) ||
+          /(opera)(?:.*version|)[ /]([\w.]+)/.exec(ua) ||
+          /(msie) ([\w.]+)/.exec(ua) ||
+          (ua.indexOf('compatible') < 0 &&
+            /(mozilla)(?:.*? rv:([\w.]+)|)/.exec(ua)) ||
+          [],
         // adding mobile flag as well
         mobile = !!userAgent.match(
           /Android|BlackBerry|iPhone|iPad|iPod|Opera Mini|IEMobile/i
@@ -221,9 +227,10 @@
     UA: undefined
   };
 
-  var glob = typeof global !== 'undefined'
-    ? global
-    : typeof window !== 'undefined'
+  var glob =
+    typeof global !== 'undefined'
+      ? global
+      : typeof window !== 'undefined'
         ? window
         : typeof WorkerGlobalScope !== 'undefined' ? self : {};
 
@@ -238,37 +245,18 @@
   glob.Konva = Konva;
   Konva.global = glob;
 
-  if (typeof exports === 'object') {
-    // runtime-check for browserify and nw.js (node-webkit)
-    if (glob.window && glob.window.document) {
-      Konva.document = glob.window.document;
-      Konva.window = glob.window;
-    } else {
-      // Node. Does not work with strict CommonJS, but
-      // only CommonJS-like enviroments that support module.exports,
-      // like Node.
-      var Canvas = require('canvas');
-      var JSDOM = require('jsdom').JSDOM;
+  Konva.document = document;
+  Konva.window = window;
 
-      Konva.window = new JSDOM(
-        '<!DOCTYPE html><html><head></head><body></body></html>'
-      ).window;
-      Konva.document = Konva.window.document;
-      Konva.window.Image = Canvas.Image;
-      Konva._nodeCanvas = Canvas;
-      Konva.isNode = true;
-    }
+  if (typeof exports === 'object') {
     module.exports = Konva;
-    return;
   } else if (typeof define === 'function' && define.amd) {
     // AMD. Register as an anonymous module.
     define(function() {
       return Konva;
     });
   }
-  Konva.document = document;
-  Konva.window = window;
-})(typeof global !== 'undefined' ? global : window);
+})();
 
 /*eslint-disable  eqeqeq, no-cond-assign, no-empty*/
 (function() {
@@ -280,7 +268,9 @@
      * @memberof Konva
      */
   Konva.Collection = function() {
-    var args = [].slice.call(arguments), length = args.length, i = 0;
+    var args = [].slice.call(arguments),
+      length = args.length,
+      i = 0;
 
     this.length = length;
     for (; i < length; i++) {
@@ -312,7 +302,9 @@
      * @memberof Konva.Collection.prototype
      */
   Konva.Collection.prototype.toArray = function() {
-    var arr = [], len = this.length, n;
+    var arr = [],
+      len = this.length,
+      n;
 
     for (n = 0; n < len; n++) {
       arr.push(this[n]);
@@ -326,7 +318,9 @@
      * @param {Array} arr
      */
   Konva.Collection.toCollection = function(arr) {
-    var collection = new Konva.Collection(), len = arr.length, n;
+    var collection = new Konva.Collection(),
+      len = arr.length,
+      n;
 
     for (n = 0; n < len; n++) {
       collection.push(arr[n]);
@@ -337,7 +331,8 @@
   // map one method by it's name
   Konva.Collection._mapMethod = function(methodName) {
     Konva.Collection.prototype[methodName] = function() {
-      var len = this.length, i;
+      var len = this.length,
+        i;
 
       var args = [].slice.call(arguments);
       for (i = 0; i < len; i++) {
@@ -792,7 +787,8 @@
          * other utils
          */
     _hasMethods: function(obj) {
-      var names = [], key;
+      var names = [],
+        key;
 
       for (key in obj) {
         if (!obj.hasOwnProperty(key)) {
@@ -816,17 +812,14 @@
       );
     },
     createCanvasElement: function() {
-      var canvas = Konva.isNode
-        ? new Konva._nodeCanvas()
-        : Konva.document.createElement('canvas');
+      var canvas = Konva.isBrowser
+        ? Konva.document.createElement('canvas')
+        : new Konva._nodeCanvas();
       // on some environments canvas.style is readonly
       try {
         canvas.style = canvas.style || {};
       } catch (e) {}
       return canvas;
-    },
-    isBrowser: function() {
-      return typeof exports !== 'object';
     },
     _isInDocument: function(el) {
       while ((el = el.parentNode)) {
@@ -837,7 +830,11 @@
       return false;
     },
     _simplifyArray: function(arr) {
-      var retArr = [], len = arr.length, util = Konva.Util, n, val;
+      var retArr = [],
+        len = arr.length,
+        util = Konva.Util,
+        n,
+        val;
 
       for (n = 0; n < len; n++) {
         val = arr[n];
@@ -1138,7 +1135,10 @@
       return [p1x, p1y, p2x, p2y];
     },
     _expandPoints: function(p, tension) {
-      var len = p.length, allPoints = [], n, cp;
+      var len = p.length,
+        allPoints = [],
+        n,
+        cp;
 
       for (n = 2; n < len - 2; n += 2) {
         cp = Konva.Util._getControlPoints(
@@ -1212,7 +1212,9 @@
           pt.x,
           pt.y
         );
-        var px = proj[0], py = proj[1], pdist = proj[2];
+        var px = proj[0],
+          py = proj[1],
+          pdist = proj[2];
         if (pdist < dist) {
           pc.x = px;
           pc.y = py;
@@ -1222,7 +1224,9 @@
       return pc;
     },
     _prepareArrayForTween: function(startArray, endArray, isClosed) {
-      var n, start = [], end = [];
+      var n,
+        start = [],
+        end = [];
       if (startArray.length > endArray.length) {
         var temp = endArray;
         endArray = startArray;
@@ -1289,8 +1293,14 @@
 (function() {
   'use strict';
   // calculate pixel ratio
-  var canvas = Konva.Util.createCanvasElement(),
-    context = canvas.getContext('2d'),
+
+  var _pixelRatio;
+  function getDevicePixelRatio() {
+    if (_pixelRatio) {
+      return _pixelRatio;
+    }
+    var canvas = Konva.Util.createCanvasElement();
+    var context = canvas.getContext('2d');
     _pixelRatio = (function() {
       var devicePixelRatio = Konva.window.devicePixelRatio || 1,
         backingStoreRatio =
@@ -1302,6 +1312,8 @@
           1;
       return devicePixelRatio / backingStoreRatio;
     })();
+    return _pixelRatio;
+  }
 
   /**
      * Canvas Renderer constructor
@@ -1327,7 +1339,8 @@
     init: function(config) {
       var conf = config || {};
 
-      var pixelRatio = conf.pixelRatio || Konva.pixelRatio || _pixelRatio;
+      var pixelRatio =
+        conf.pixelRatio || Konva.pixelRatio || getDevicePixelRatio();
 
       this.pixelRatio = pixelRatio;
       this._canvas = Konva.Util.createCanvasElement();
@@ -1390,7 +1403,8 @@
       this.width = this._canvas.width = width * this.pixelRatio;
       this._canvas.style.width = width + 'px';
 
-      var pixelRatio = this.pixelRatio, _context = this.getContext()._context;
+      var pixelRatio = this.pixelRatio,
+        _context = this.getContext()._context;
       _context.scale(pixelRatio, pixelRatio);
     },
     /**
@@ -1403,7 +1417,8 @@
       // take into account pixel ratio
       this.height = this._canvas.height = height * this.pixelRatio;
       this._canvas.style.height = height + 'px';
-      var pixelRatio = this.pixelRatio, _context = this.getContext()._context;
+      var pixelRatio = this.pixelRatio,
+        _context = this.getContext()._context;
       _context.scale(pixelRatio, pixelRatio);
     },
     /**
@@ -1461,7 +1476,8 @@
 
   Konva.SceneCanvas = function(config) {
     var conf = config || {};
-    var width = conf.width || 0, height = conf.height || 0;
+    var width = conf.width || 0,
+      height = conf.height || 0;
 
     Konva.Canvas.call(this, conf);
     this.context = new Konva.SceneContext(this);
@@ -1472,7 +1488,8 @@
 
   Konva.HitCanvas = function(config) {
     var conf = config || {};
-    var width = conf.width || 0, height = conf.height || 0;
+    var width = conf.width || 0,
+      height = conf.height || 0;
 
     Konva.Canvas.call(this, conf);
     this.context = new Konva.HitContext(this);
@@ -2514,7 +2531,10 @@
         */
     cache: function(config) {
       var conf = config || {},
-        rect = this.getClientRect(true),
+        rect = this.getClientRect({
+          skipTransform: true,
+          relativeTo: this.getParent()
+        }),
         width = conf.width || rect.width,
         height = conf.height || rect.height,
         pixelRatio = conf.pixelRatio,
@@ -2524,7 +2544,13 @@
         drawBorder = conf.drawBorder || false;
 
       if (!width || !height) {
-        throw new Error('Width or height of caching configuration equals 0.');
+        // make throw async, because we don't need to stop funcion
+        setTimeout(function() {
+          Konva.Util.throw(
+            'Width or height of caching configuration equals 0. Caching is ignored.'
+          );
+        });
+        return;
       }
 
       width += offset * 2;
@@ -2534,10 +2560,10 @@
       y -= offset;
 
       var cachedSceneCanvas = new Konva.SceneCanvas({
-        pixelRatio: pixelRatio,
-        width: width,
-        height: height
-      }),
+          pixelRatio: pixelRatio,
+          width: width,
+          height: height
+        }),
         cachedFilterCanvas = new Konva.SceneCanvas({
           pixelRatio: pixelRatio,
           width: width,
@@ -2601,7 +2627,9 @@
          * The rectangle position is relative to parent container.
          * @method
          * @memberof Konva.Node.prototype
-         * @param {Boolean} [skipTransform] flag should we skip transformation to rectangle
+         * @param {Object} config
+         * @param {Boolean} [config.skipTransform] should we apply transform to node for calculating rect?
+         * @param {Object} [config.relativeTo] calculate client rect relative to one of the parents
          * @returns {Object} rect with {x, y, width, height} properties
          * @example
          * var rect = new Konva.Rect({
@@ -2616,7 +2644,7 @@
          * });
          *
          * // get client rect without think off transformations (position, rotation, scale, offset, etc)
-         * rect.getClientRect(true);
+         * rect.getClientRect({ skipTransform: true});
          * // returns {
          * //     x : -2,   // two pixels for stroke / 2
          * //     y : -2,
@@ -2633,7 +2661,7 @@
       // redefine in Container and Shape
       throw new Error('abstract "getClientRect" method call');
     },
-    _transformedRect: function(rect) {
+    _transformedRect: function(rect, top) {
       var points = [
         { x: rect.x, y: rect.y },
         { x: rect.x + rect.width, y: rect.y },
@@ -2641,7 +2669,7 @@
         { x: rect.x, y: rect.y + rect.height }
       ];
       var minX, minY, maxX, maxY;
-      var trans = this.getTransform();
+      var trans = this.getAbsoluteTransform(top);
       points.forEach(function(point) {
         var transformed = trans.point(point);
         if (minX === undefined) {
@@ -2679,7 +2707,8 @@
       context.restore();
     },
     _drawCachedHitCanvas: function(context) {
-      var cachedCanvas = this._cache.canvas, hitCanvas = cachedCanvas.hit;
+      var cachedCanvas = this._cache.canvas,
+        hitCanvas = cachedCanvas.hit;
       context.save();
       context.translate(this._cache.canvas.x, this._cache.canvas.y);
       context.drawImage(hitCanvas._canvas, 0, 0);
@@ -3007,7 +3036,8 @@
         * })
         */
     getAncestors: function() {
-      var parent = this.getParent(), ancestors = new Konva.Collection();
+      var parent = this.getParent(),
+        ancestors = new Konva.Collection();
 
       while (parent) {
         ancestors.push(parent);
@@ -3081,7 +3111,8 @@
       return this._getCache(LISTENING, this._isListening);
     },
     _isListening: function() {
-      var listening = this.getListening(), parent = this.getParent();
+      var listening = this.getListening(),
+        parent = this.getParent();
 
       // the following conditions are a simplification of the truth table above.
       // please modify carefully
@@ -3118,7 +3149,8 @@
       return this._getCache(VISIBLE, this._isVisible);
     },
     _isVisible: function() {
-      var visible = this.getVisible(), parent = this.getParent();
+      var visible = this.getVisible(),
+        parent = this.getParent();
 
       // the following conditions are a simplification of the truth table above.
       // please modify carefully
@@ -3186,7 +3218,13 @@
          * @returns {Integer}
          */
     getAbsoluteZIndex: function() {
-      var depth = this.getDepth(), that = this, index = 0, nodes, len, n, child;
+      var depth = this.getDepth(),
+        that = this,
+        index = 0,
+        nodes,
+        len,
+        n,
+        child;
 
       function addChildren(children) {
         nodes = [];
@@ -3223,7 +3261,8 @@
          * @returns {Integer}
          */
     getDepth: function() {
-      var depth = 0, parent = this.parent;
+      var depth = 0,
+        parent = this.parent;
 
       while (parent) {
         depth++;
@@ -3271,7 +3310,8 @@
          * @returns {Konva.Node}
          */
     setAbsolutePosition: function(pos) {
-      var origTrans = this._clearTransform(), it;
+      var origTrans = this._clearTransform(),
+        it;
 
       // don't clear translation
       this.attrs.x = origTrans.x;
@@ -3366,7 +3406,10 @@
       return this;
     },
     _eachAncestorReverse: function(func, top) {
-      var family = [], parent = this.getParent(), len, n;
+      var family = [],
+        parent = this.getParent(),
+        len,
+        n;
 
       // if top node is defined, and this node is top node,
       // there's no need to build a family tree.  just execute
@@ -3427,7 +3470,8 @@
         Konva.Util.warn('Node has no parent. moveUp function is ignored.');
         return false;
       }
-      var index = this.index, len = this.parent.getChildren().length;
+      var index = this.index,
+        len = this.parent.getChildren().length;
       if (index < len - 1) {
         this.parent.children.splice(index, 1);
         this.parent.children.splice(index + 1, 0, this);
@@ -3540,7 +3584,12 @@
          * @returns {Object}
          */
     toObject: function() {
-      var obj = {}, attrs = this.getAttrs(), key, val, getter, defaultValue;
+      var obj = {},
+        attrs = this.getAttrs(),
+        key,
+        val,
+        getter,
+        defaultValue;
 
       obj.attrs = {};
 
@@ -3742,7 +3791,9 @@
       }
     },
     _getAbsoluteTransform: function(top) {
-      var at = new Konva.Transform(), transformsEnabled, trans;
+      var at = new Konva.Transform(),
+        transformsEnabled,
+        trans;
 
       // start with stage and traverse downwards to self
       this._eachAncestorReverse(function(node) {
@@ -3783,7 +3834,8 @@
         parent = parent.getParent();
       }
 
-      var scaleX = 1, scaleY = 1;
+      var scaleX = 1,
+        scaleY = 1;
 
       // start with stage and traverse downwards to self
       this._eachAncestorReverse(function(node) {
@@ -3899,10 +3951,10 @@
         y = config.y || 0,
         pixelRatio = config.pixelRatio || 1,
         canvas = new Konva.SceneCanvas({
-          width: config.width ||
-            this.getWidth() ||
-            (stage ? stage.getWidth() : 0),
-          height: config.height ||
+          width:
+            config.width || this.getWidth() || (stage ? stage.getWidth() : 0),
+          height:
+            config.height ||
             this.getHeight() ||
             (stage ? stage.getHeight() : 0),
           pixelRatio: pixelRatio
@@ -3958,7 +4010,8 @@
          */
     toDataURL: function(config) {
       config = config || {};
-      var mimeType = config.mimeType || null, quality = config.quality || null;
+      var mimeType = config.mimeType || null,
+        quality = config.quality || null;
       return this._toKonvaCanvas(config).toDataURL(mimeType, quality);
     },
     /**
@@ -4045,7 +4098,9 @@
         : [];
     },
     _off: function(type, name) {
-      var evtListeners = this.eventListeners[type], i, evtName;
+      var evtListeners = this.eventListeners[type],
+        i,
+        evtName;
 
       for (i = 0; i < evtListeners.length; i++) {
         evtName = evtListeners[i].name;
@@ -4170,7 +4225,8 @@
          * node.setAttr('x', 5);
          */
     setAttr: function(attr, val) {
-      var method = SET + Konva.Util._capitalize(attr), func = this[method];
+      var method = SET + Konva.Util._capitalize(attr),
+        func = this[method];
 
       if (Konva.Util._isFunction(func)) {
         func.call(this, val);
@@ -4259,7 +4315,8 @@
       }
     },
     _fire: function(eventType, evt) {
-      var events = this.eventListeners[eventType], i;
+      var events = this.eventListeners[eventType],
+        i;
 
       evt = evt || {};
       evt.currentTarget = this;
@@ -4875,7 +4932,7 @@
   };
 })();
 
-(function() {
+(function(Konva) {
   'use strict';
   /**
      * Brighten Filter.
@@ -4919,7 +4976,7 @@
     * @param {Number} brightness value between -1 and 1
     * @returns {Number}
     */
-})();
+})(Konva);
 
 (function() {
   'use strict';
@@ -4950,7 +5007,7 @@
  the Gauss filter
  master repo: https://github.com/pavelpower/kineticjsGaussFilter
 */
-(function() {
+(function(Konva) {
   'use strict';
   /*
 
@@ -5799,7 +5856,7 @@
     * @param {Integer} radius
     * @returns {Integer}
     */
-})();
+})(Konva);
 
 /*eslint-disable  max-depth */
 (function() {
@@ -7196,18 +7253,24 @@
       a = 0;
 
     // Find the largest radius
-    var rad, rMax = Math.sqrt(xMid * xMid + yMid * yMid);
+    var rad,
+      rMax = Math.sqrt(xMid * xMid + yMid * yMid);
     x = xSize - xMid;
     y = ySize - yMid;
     rad = Math.sqrt(x * x + y * y);
     rMax = rad > rMax ? rad : rMax;
 
     // We'll be uisng y as the radius, and x as the angle (theta=t)
-    var rSize = ySize, tSize = xSize, radius, theta;
+    var rSize = ySize,
+      tSize = xSize,
+      radius,
+      theta;
 
     // We want to cover all angles (0-360) and we need to convert to
     // radians (*PI/180)
-    var conversion = 360 / tSize * Math.PI / 180, sin, cos;
+    var conversion = 360 / tSize * Math.PI / 180,
+      sin,
+      cos;
 
     // var x1, x2, x1i, x2i, y1, y2, y1i, y2i, scale;
 
@@ -7269,7 +7332,8 @@
       a = 0;
 
     // Find the largest radius
-    var rad, rMax = Math.sqrt(xMid * xMid + yMid * yMid);
+    var rad,
+      rMax = Math.sqrt(xMid * xMid + yMid * yMid);
     x = xSize - xMid;
     y = ySize - yMid;
     rad = Math.sqrt(x * x + y * y);
@@ -7318,7 +7382,6 @@
   //Konva.Filters.FromPolar = Konva.Util._FilterWrapDoubleBuffer(FromPolar);
 
   // create a temporary canvas for working - shared between multiple calls
-  var tempCanvas = Konva.Util.createCanvasElement();
 
   /*
      * Kaleidoscope Filter.
@@ -7333,7 +7396,8 @@
      * node.kaleidoscopeAngle(45);
      */
   Konva.Filters.Kaleidoscope = function(imageData) {
-    var xSize = imageData.width, ySize = imageData.height;
+    var xSize = imageData.width,
+      ySize = imageData.height;
 
     var x, y, xoff, i, r, g, b, a, srcPos, dstPos;
     var power = Math.round(this.kaleidoscopePower());
@@ -7345,6 +7409,7 @@
     }
 
     // Work with our shared buffer canvas
+    var tempCanvas = Konva.Util.createCanvasElement();
     tempCanvas.width = xSize;
     tempCanvas.height = ySize;
     var scratchData = tempCanvas
@@ -7370,7 +7435,9 @@
     // Copy the offset region to 0
     // Depending on the size of filter and location of the offset we may need
     // to copy the section backwards to prevent it from rewriting itself
-    var xStart = 0, xEnd = sectionSize, xDelta = 1;
+    var xStart = 0,
+      xEnd = sectionSize,
+      xDelta = 1;
     if (offset + minSectionSize > xSize) {
       xStart = sectionSize;
       xEnd = 0;
@@ -7898,7 +7965,11 @@
           !layerUnderDrag)
       );
     },
-    getClientRect: function(skipTransform) {
+    getClientRect: function(attrs) {
+      attrs = attrs || {};
+      var skipTransform = attrs.skipTransform;
+      var relativeTo = attrs.relativeTo;
+
       var minX, minY, maxX, maxY;
       var selfRect = {
         x: 0,
@@ -7906,12 +7977,13 @@
         width: 0,
         height: 0
       };
+      var that = this;
       this.children.each(function(child) {
         // skip invisible children
         if (!child.isVisible()) {
           return;
         }
-        var rect = child.getClientRect();
+        var rect = child.getClientRect({ relativeTo: that });
 
         // skip invisible children (like empty groups)
         // or don't skip... hmmm...
@@ -7943,7 +8015,7 @@
       }
 
       if (!skipTransform) {
-        return this._transformedRect(selfRect);
+        return this._transformedRect(selfRect, relativeTo);
       }
       return selfRect;
     }
@@ -8379,7 +8451,11 @@
         height: size.height
       };
     },
-    getClientRect: function(skipTransform) {
+    getClientRect: function(attrs) {
+      attrs = attrs || {};
+      var skipTransform = attrs.skipTransform;
+      var relativeTo = attrs.relativeTo;
+
       var fillRect = this.getSelfRect();
 
       var strokeWidth = (this.hasStroke() && this.strokeWidth()) || 0;
@@ -8415,7 +8491,7 @@
           fillRect.y
       };
       if (!skipTransform) {
-        return this._transformedRect(rect);
+        return this._transformedRect(rect, relativeTo);
       }
       return rect;
     },
@@ -9938,7 +10014,9 @@
          * @memberof Konva.Stage.prototype
          */
     clear: function() {
-      var layers = this.children, len = layers.length, n;
+      var layers = this.children,
+        len = layers.length,
+        n;
 
       for (n = 0; n < len; n++) {
         layers[n].clear();
@@ -10147,7 +10225,10 @@
 
       // draw layer and append canvas to container
       layer.draw();
-      this.content.appendChild(layer.canvas._canvas);
+
+      if (Konva.isBrowser) {
+        this.content.appendChild(layer.canvas._canvas);
+      }
 
       // chainable
       return this;
@@ -10167,6 +10248,9 @@
       return this.getChildren();
     },
     _bindContentEvents: function() {
+      if (!Konva.isBrowser) {
+        return;
+      }
       for (var n = 0; n < eventsLength; n++) {
         addEvent(this, EVENTS[n]);
       }
@@ -10421,7 +10505,8 @@
     },
     _touchmove: function(evt) {
       this._setPointerPosition(evt);
-      var dd = Konva.DD, shape;
+      var dd = Konva.DD,
+        shape;
       if (!Konva.isDragging()) {
         shape = this.getIntersection(this.getPointerPosition());
         if (shape && shape.isListening()) {
@@ -10453,7 +10538,9 @@
       this._fire(CONTENT_WHEEL, { evt: evt });
     },
     _setPointerPosition: function(evt) {
-      var contentPosition = this._getContentPosition(), x = null, y = null;
+      var contentPosition = this._getContentPosition(),
+        x = null,
+        y = null;
       evt = evt ? evt : window.event;
 
       // touch events
@@ -10487,14 +10574,18 @@
       };
     },
     _buildDOM: function() {
+      // the buffer canvas pixel ratio must be 1 because it is used as an
+      // intermediate canvas before copying the result onto a scene canvas.
+      // not setting it to 1 will result in an over compensation
+      this.bufferCanvas = new Konva.SceneCanvas();
+      this.bufferHitCanvas = new Konva.HitCanvas({ pixelRatio: 1 });
+
+      if (!Konva.isBrowser) {
+        return;
+      }
       var container = this.getContainer();
       if (!container) {
-        if (Konva.Util.isBrowser()) {
-          throw 'Stage has no container. A container is required.';
-        } else {
-          // automatically create element for jsdom in nodejs env
-          container = Konva.document.createElement(DIV);
-        }
+        throw 'Stage has no container. A container is required.';
       }
       // clear content inside container
       container.innerHTML = EMPTY_STRING;
@@ -10504,18 +10595,16 @@
       this.content.style.position = RELATIVE;
       this.content.className = KONVA_CONTENT;
       this.content.setAttribute('role', 'presentation');
-      container.appendChild(this.content);
 
-      // the buffer canvas pixel ratio must be 1 because it is used as an
-      // intermediate canvas before copying the result onto a scene canvas.
-      // not setting it to 1 will result in an over compensation
-      this.bufferCanvas = new Konva.SceneCanvas();
-      this.bufferHitCanvas = new Konva.HitCanvas({ pixelRatio: 1 });
+      container.appendChild(this.content);
 
       this._resizeDOM();
     },
     _onContent: function(typesStr, handler) {
-      var types = typesStr.split(SPACE), len = types.length, n, baseEvent;
+      var types = typesStr.split(SPACE),
+        len = types.length,
+        n,
+        baseEvent;
 
       for (n = 0; n < len; n++) {
         baseEvent = types[n];
@@ -10553,7 +10642,7 @@
      */
 })();
 
-(function() {
+(function(Konva) {
   'use strict';
   /**
      * BaseLayer constructor.
@@ -10590,8 +10679,6 @@
      * @param {Number} [config.clipHeight] set clip height
      * @param {Function} [config.clipFunc] set clip func
 
-     * @example
-     * var layer = new Konva.Layer();
      */
   Konva.BaseLayer = function(config) {
     this.___init(config);
@@ -10606,47 +10693,47 @@
       return this.canvas._canvas.createPNGStream();
     },
     /**
-         * get layer canvas
-         * @method
-         * @memberof Konva.BaseLayer.prototype
-         */
+     * get layer canvas
+     * @method
+     * @memberof Konva.BaseLayer.prototype
+     */
     getCanvas: function() {
       return this.canvas;
     },
     /**
-         * get layer hit canvas
-         * @method
-         * @memberof Konva.BaseLayer.prototype
-         */
+     * get layer hit canvas
+     * @method
+     * @memberof Konva.BaseLayer.prototype
+     */
     getHitCanvas: function() {
       return this.hitCanvas;
     },
     /**
-         * get layer canvas context
-         * @method
-         * @memberof Konva.BaseLayer.prototype
-         */
+     * get layer canvas context
+     * @method
+     * @memberof Konva.BaseLayer.prototype
+     */
     getContext: function() {
       return this.getCanvas().getContext();
     },
     /**
-         * clear scene and hit canvas contexts tied to the layer
-         * @method
-         * @memberof Konva.BaseLayer.prototype
-         * @param {Object} [bounds]
-         * @param {Number} [bounds.x]
-         * @param {Number} [bounds.y]
-         * @param {Number} [bounds.width]
-         * @param {Number} [bounds.height]
-         * @example
-         * layer.clear();
-         * layer.clear({
-         *   x : 0,
-         *   y : 0,
-         *   width : 100,
-         *   height : 100
-         * });
-         */
+     * clear scene and hit canvas contexts tied to the layer
+     * @method
+     * @memberof Konva.BaseLayer.prototype
+     * @param {Object} [bounds]
+     * @param {Number} [bounds.x]
+     * @param {Number} [bounds.y]
+     * @param {Number} [bounds.width]
+     * @param {Number} [bounds.height]
+     * @example
+     * layer.clear();
+     * layer.clear({
+     *   x : 0,
+     *   y : 0,
+     *   width : 100,
+     *   height : 100
+     * });
+     */
     clear: function(bounds) {
       this.getContext().clear(bounds);
       return this;
@@ -10755,15 +10842,15 @@
       return this;
     },
     /**
-         * get/set width of layer.getter return width of stage. setter doing nothing.
-         * if you want change width use `stage.width(value);`
-         * @name width
-         * @method
-         * @memberof Konva.BaseLayer.prototype
-         * @returns {Number}
-         * @example
-         * var width = layer.width();
-         */
+     * get/set width of layer.getter return width of stage. setter doing nothing.
+     * if you want change width use `stage.width(value);`
+     * @name width
+     * @method
+     * @memberof Konva.BaseLayer.prototype
+     * @returns {Number}
+     * @example
+     * var width = layer.width();
+     */
     getWidth: function() {
       if (this.parent) {
         return this.parent.getWidth();
@@ -10775,15 +10862,15 @@
       );
     },
     /**
-         * get/set height of layer.getter return height of stage. setter doing nothing.
-         * if you want change height use `stage.height(value);`
-         * @name height
-         * @method
-         * @memberof Konva.BaseLayer.prototype
-         * @returns {Number}
-         * @example
-         * var height = layer.height();
-         */
+     * get/set height of layer.getter return height of stage. setter doing nothing.
+     * if you want change height use `stage.height(value);`
+     * @name height
+     * @method
+     * @memberof Konva.BaseLayer.prototype
+     * @returns {Number}
+     * @example
+     * var height = layer.height();
+     */
     getHeight: function() {
       if (this.parent) {
         return this.parent.getHeight();
@@ -10807,26 +10894,26 @@
   // add getters and setters
   Konva.Factory.addGetterSetter(Konva.BaseLayer, 'clearBeforeDraw', true);
   /**
-     * get/set clearBeforeDraw flag which determines if the layer is cleared or not
-     *  before drawing
-     * @name clearBeforeDraw
-     * @method
-     * @memberof Konva.BaseLayer.prototype
-     * @param {Boolean} clearBeforeDraw
-     * @returns {Boolean}
-     * @example
-     * // get clearBeforeDraw flag
-     * var clearBeforeDraw = layer.clearBeforeDraw();
-     *
-     * // disable clear before draw
-     * layer.clearBeforeDraw(false);
-     *
-     * // enable clear before draw
-     * layer.clearBeforeDraw(true);
-     */
+   * get/set clearBeforeDraw flag which determines if the layer is cleared or not
+   *  before drawing
+   * @name clearBeforeDraw
+   * @method
+   * @memberof Konva.BaseLayer.prototype
+   * @param {Boolean} clearBeforeDraw
+   * @returns {Boolean}
+   * @example
+   * // get clearBeforeDraw flag
+   * var clearBeforeDraw = layer.clearBeforeDraw();
+   *
+   * // disable clear before draw
+   * layer.clearBeforeDraw(false);
+   *
+   * // enable clear before draw
+   * layer.clearBeforeDraw(true);
+   */
 
   Konva.Collection.mapMethods(Konva.BaseLayer);
-})();
+})(Konva);
 
 (function() {
   'use strict';
@@ -12318,11 +12405,17 @@
 
     // methods
     _drag: function(evt) {
-      var dd = Konva.DD, node = dd.node;
-
+      var dd = Konva.DD,
+        node = dd.node;
       if (node) {
         if (!dd.isDragging) {
           var pos = node.getStage().getPointerPosition();
+          // it is possible that pos is undefined
+          // reattach it
+          if (!pos) {
+            node.getStage()._setPointerPosition(evt);
+            pos = node.getStage().getPointerPosition();
+          }
           var dragDistance = node.dragDistance();
           var distance = Math.max(
             Math.abs(pos.x - dd.startPointerPos.x),
@@ -12361,7 +12454,9 @@
       }
     },
     _endDragBefore: function(evt) {
-      var dd = Konva.DD, node = dd.node, layer;
+      var dd = Konva.DD,
+        node = dd.node,
+        layer;
 
       if (node) {
         layer = node.getLayer();
@@ -12468,7 +12563,8 @@
      * @memberof Konva.Node.prototype
      */
   Konva.Node.prototype.stopDrag = function() {
-    var dd = Konva.DD, evt = {};
+    var dd = Konva.DD,
+      evt = {};
     dd._endDragBefore(evt);
     dd._endDragAfter(evt);
   };
@@ -12606,15 +12702,17 @@
      * node.draggable(false);
      */
 
-  var html = Konva.document.documentElement;
-  html.addEventListener('mouseup', Konva.DD._endDragBefore, true);
-  html.addEventListener('touchend', Konva.DD._endDragBefore, true);
+  if (Konva.isBrowser) {
+    var html = Konva.document.documentElement;
+    html.addEventListener('mouseup', Konva.DD._endDragBefore, true);
+    html.addEventListener('touchend', Konva.DD._endDragBefore, true);
 
-  html.addEventListener('mousemove', Konva.DD._drag);
-  html.addEventListener('touchmove', Konva.DD._drag);
+    html.addEventListener('mousemove', Konva.DD._drag);
+    html.addEventListener('touchmove', Konva.DD._drag);
 
-  html.addEventListener('mouseup', Konva.DD._endDragAfter, false);
-  html.addEventListener('touchend', Konva.DD._endDragAfter, false);
+    html.addEventListener('mouseup', Konva.DD._endDragAfter, false);
+    html.addEventListener('touchend', Konva.DD._endDragAfter, false);
+  }
 })();
 
 (function() {
@@ -12793,7 +12891,7 @@
   Konva.Collection.mapMethods(Konva.Rect);
 })();
 
-(function() {
+(function(Konva) {
   'use strict';
   // the 0.0001 offset fixes a bug in Chrome 27
   var PIx2 = Math.PI * 2 - 0.0001, CIRCLE = 'Circle';
@@ -12948,7 +13046,7 @@
      */
 
   Konva.Collection.mapMethods(Konva.Circle);
-})();
+})(Konva);
 
 (function() {
   'use strict';
@@ -13536,7 +13634,7 @@
   Konva.Collection.mapMethods(Konva.Wedge);
 })();
 
-(function() {
+(function(Konva) {
   'use strict';
   /**
      * Arc constructor
@@ -13749,7 +13847,7 @@
      */
 
   Konva.Collection.mapMethods(Konva.Arc);
-})();
+})(Konva);
 
 (function() {
   'use strict';
@@ -14104,8 +14202,15 @@
       'letterSpacing'
     ],
     // cached variables
-    attrChangeListLen = ATTR_CHANGE_LIST.length,
+    attrChangeListLen = ATTR_CHANGE_LIST.length;
+  var dummyContext;
+  function getDummyContext() {
+    if (dummyContext) {
+      return dummyContext;
+    }
     dummyContext = Konva.Util.createCanvasElement().getContext(CONTEXT_2D);
+    return dummyContext;
+  }
 
   /**
      * Text constructor
@@ -14276,7 +14381,9 @@
 
       // draw text lines
       for (n = 0; n < textArrLen; n++) {
-        var obj = textArr[n], text = obj.text, width = obj.width;
+        var obj = textArr[n],
+          text = obj.text,
+          width = obj.width;
 
         // horizontal alignment
         context.save();
@@ -14338,7 +14445,8 @@
       context.restore();
     },
     _hitFunc: function(context) {
-      var width = this.getWidth(), height = this.getHeight();
+      var width = this.getWidth(),
+        height = this.getHeight();
 
       context.beginPath();
       context.rect(0, 0, width, height);
@@ -14381,7 +14489,7 @@
         this.attrs.height === AUTO || this.attrs.height === undefined;
       return isAuto
         ? this.getTextHeight() * this.textArr.length * this.getLineHeight() +
-            this.getPadding() * 2
+          this.getPadding() * 2
         : this.attrs.height;
     },
     /**
@@ -14403,7 +14511,9 @@
       return this.textHeight;
     },
     _getTextSize: function(text) {
-      var _context = dummyContext, fontSize = this.getFontSize(), metrics;
+      var _context = getDummyContext(),
+        fontSize = this.getFontSize(),
+        metrics;
 
       _context.save();
       _context.font = this._getContextFont();
@@ -14450,7 +14560,7 @@
       var latterSpacing = this.getLetterSpacing();
       var length = text.length;
       return (
-        dummyContext.measureText(text).width +
+        getDummyContext().measureText(text).width +
         (length ? latterSpacing * (length - 1) : 0)
       );
     },
@@ -14472,8 +14582,8 @@
         wrapAtWord = wrap !== CHAR && shouldWrap;
 
       this.textArr = [];
-      dummyContext.save();
-      dummyContext.font = this._getContextFont();
+      getDummyContext().save();
+      getDummyContext().font = this._getContextFont();
       for (var i = 0, max = lines.length; i < max; ++i) {
         var line = lines[i];
 
@@ -14488,7 +14598,10 @@
                          * use binary search to find the longest substring that
                          * that would fit in the specified width
                          */
-            var low = 0, high = line.length, match = '', matchWidth = 0;
+            var low = 0,
+              high = line.length,
+              match = '',
+              matchWidth = 0;
             while (low < high) {
               var mid = (low + high) >>> 1,
                 substr = line.slice(0, mid + 1),
@@ -14561,7 +14674,7 @@
           break;
         }
       }
-      dummyContext.restore();
+      getDummyContext().restore();
       this.textHeight = fontSize;
       // var maxTextWidth = 0;
       // for(var j = 0; j < this.textArr.length; j++) {
@@ -17909,7 +18022,7 @@
   Konva.Collection.mapMethods(Konva.Tag);
 })();
 
-(function() {
+(function(Konva) {
   'use strict';
   /**
      * Arrow constructor
@@ -18102,4 +18215,4 @@
 
   Konva.Factory.addGetterSetter(Konva.Arrow, 'pointerAtBeginning', false);
   Konva.Collection.mapMethods(Konva.Arrow);
-})();
+})(Konva);